move 'measurement_dir' fixture to 'tests/conftest' dir
